pyproject.toml
[tool.poetry]
name = "textcase"
version = "0.1.0"
description = "A full-stack text-based CASE tool"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "textcase", from = "src"}]
license = "MIT"
include = ["textcase/py.typed"]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
pyyaml = "^6.0.1"
uuid = "^1.30"

[tool.poetry.scripts]
tse = "textcase.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "textcase"
version = "0.1.0"
description = "A full-stack text-based CASE tool"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]

[project.scripts]
tse = "textcase.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.0"
flake8 = "^6.1.0"
